Move spotify logic to its own client
